explorer/requirements.txt
streamlit==1.39.0
pandas==2.2.2
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
